make validation work (dirty way) + cleanup
